Upgrade to Exposed 0.22.1;

+++ scp-users/src/main/kotlin/io/github/tiscs/scp/models/Models.kt
@@ -18,9 +18,9 @@ data class Event<T>(
         val model: T
 )
 
-data class Page<T>(val total: Int, val page: Int, val size: Int, val items: List<T>) {
+data class Page<T>(val total: Long, val page: Int, val size: Int, val items: List<T>) {
     constructor(rows: SizedIterable<ResultRow>, page: Int, size: Int, mapper: (ResultRow) -> T, countOnly: Boolean = false) :
-            this(rows.count(), page, size, if (countOnly) emptyList() else rows.limit(size, page * size).map(mapper))
+            this(rows.count(), page, size, if (countOnly) emptyList() else rows.limit(size, page.toLong() * size).map(mapper))
 }
 
 enum class Gender {
